requirements.txt
psycopg2-binary==2.9.7
pybit==5.5.0
numpy==1.25.2
Ta-Lib==0.4.18
python-dotenv==1.0.0
plotly==5.16.1
pandas==2.1.0
python-coinmarketcap==0.5
schedule==1.2.0
redis==5.0.0

